fixed delta bug
we had a bug where only the first ball would change significantly position, even though all of them weresupposed to move the same. The problem was in the fact that getDelta() was withing each object's update() method, thus making the variations very small for all but the first object.
Now delta is calculated in the animate() method and passed as an argument like this: object.update(delta)