implemented colision detection on ball placement
It was required to change some functions which "detected colisions". They were in fact manipulating values. But when creating the balls, we only really know if they colide with existing balls.

added another function "detect_movable_colision" in colisions.js that detects if a ball is clashing with another one. This doesn't feel right, but it works